home-manager: move install command into main tool
This simplifies things slightly since we can use the code
infrastructure that is already present in the home-manager tool.

We also add an install option `--with-flake` which causes an initial
`flake.nix` file to be generated.

Finally, we update the installation instructions for the Nix Flakes
standalone setup to use the new install sub-command.